requirements.txt
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.62.0
opencv-python>=4.5.0